Rename message_handlers to channels
